py002django/requirements.txt
Django==4.0.6
cryptography==37.0.4
mysqlclient==2.1.1
mysql-connector-python==8.0.29
Pillow==9.2.0
djangorestframework==3.14.0
